+++ src/utils/formatters/data.js
@@ -7,6 +7,7 @@ import {
   converters,
 } from "@umb-network/toolbox";
 import int64 from "int64-buffer";
+import { formatDistanceToNowStrict } from "date-fns";
 
 export const arrayToReadableJSON = (array) => {
   const formattedArray = array.map((element) => `"${element}"`);
@@ -102,4 +103,15 @@ export function textToHex(text) {
   } catch {
     return utils.hexlify(utils.toUtf8Bytes(text));
   }
+}
+
+export function formatTimestamp(timestamp) {
+  return new Date(timestamp)
+    .toISOString()
+    .replace("T", " ")
+    .replace(".000", " ");
+}
+
+export function formatTimestampAge(timestamp) {
+  return `${formatDistanceToNowStrict(new Date(timestamp))} ago`;
 }